Update start.sh
added if statements to the removing of x11 folders and creation of mods folder. this will help not creating a new dir over the old one possibly deleting the old Mods folder with active mods in it.

added an echo statement at line 8 to also inform the user that the persistentdata folder has changed to persistent_data. this was an issue i had that i have to investigate further upon entering a new world when update 1.3 was pushed.